// frontend/src/app/api-gen/models/compile-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { StatementResultInfo } from '../models/statement-result-info';
export interface CompileResponse {
  code_b?: (string | null);
  env_a: string;
  env_b: string;
  err_count?: number;
  notes?: Array<string>;
  object_name: string;
  object_type: string;
  ok_count?: number;
  results?: Array<StatementResultInfo>;
  schema_name: string;
  script?: (string | null);
  status: string;
}